[port] Fixed up prvSetupTimerInterrupt
InitInterruptController() is called in the beginning of main already. This
caused a lot of confusion for me as I thought this was the only place and
I just couldn't understand why my interrupt vector was overwritten.

RegisterInterrupt needs to be called _after_ the timer registers are set,
because RegisterInterrupt activates interrupts after it has set the vector.