interrupts: functions return void
Previously, functions returned a status based only on the input to the
function. Nothing inside the function itself could fail, or influence
the outcome of the return value. This made little sense: if the caller
is checking for errors, the only errors he will find are from his own
invalid input! Leave the burden of error checking to the caller. Still
check for out-of-range errors, but ignore them.